[DF-10434] Build fix
Former-commit-id: 395786bd6a7fe3591bc3739918bc1b7b7843fb52

+++ Sources/Internal/UI/Private/CEFController.cpp
@@ -178,7 +178,7 @@ void CEFControllerImpl::CleanCache()
     FilePath cachePath = GetCachePath();
     Vector<FilePath> cacheDirContent = fs->EnumerateFilesInDirectory(cachePath);
 
-    size_t cacheSize = 0;
+    uint32 cacheSize = 0;
     for (const FilePath& path : cacheDirContent)
     {
         uint32 size = 0;
